move GSubst into a separate module
